+++ orchestrator/src/worker/controller/mod.rs
@@ -1,3 +1,4 @@
+pub mod delivery_retry;
 pub mod event_worker;
 pub mod priority_queue_worker;
 pub mod worker_controller;
